Add community edition
I'm using the community edition now. That does not mean that I want to
install it locally. Add a docker version of the community edition as
well.

A thing to note here is that CE store the config in a different place
(or if it's the newest version). So a new folder is needed.