Move location of Adwaita cursors
While the icons in the Adwaita theme are nice, the cursors are horrible.
Bloated things that are massive. I much prefer the standard cursors that
come with X. A simple way to force this is to simple move the
cursor-folder in the Adwaita theme to somewhere the theme can't find
them. So that's what I do. This results in everything but the cursors in
the Adwaita theme to be loaded. Very nice.