Add OpenSCAD
Running OpenSCAD in a container, cause why not. I have chosen to not
pull in pulse-audio connection as I don't find it super helpfull to have
audio bells for things like this. It can be added without trouble, check
the chromium image for how to to it.

libGL throws an error about nod being able to open /dev/dri/card0 and
loading drivers. Everything works though so I haven't bothered with it
for now.

It fails to open the image specified on the command-line, but you can
open it from within the editor so meh. Would be nice if it worked but
I'm not sure where it fails to do the right thing.