Add jerns (jdegenstein) fork of cq-editor
The fork can be found here https://github.com/jdegenstein/jmwright-CQ-Editor
This fork has some sane defaults for the mouse (lmb to rotate, mmb to
zoom and rmb to pan/strafe) as well as being darkmode. Easier on the
eyes.
Not only that, it includes a lot of the common libs many are using, like
cq_gears, cq_cache, cq_more, cq_warehouse, and build123d. That's a lot
to get by default.