Including docker-on-x
Since the GitLab repo now requires login and paddy-hack no longer hosts
the blog post under that link, and I can't find it elsewhere, I have
decided to include the docker-on-x script in this repository to make it
easier to find and use. The copyrights stays the same. The script is
GPL-3 so is safe to redistribute.